now velocity of player depends on dt; added acceleration and deceleration of time

+++ game_sources/include/GameField/HealthPack.h
@@ -12,7 +12,7 @@ public:
 
 	shared_ptr<PlayField> GetPlayfield() const { return m_playfield.lock(); }
 
-	void Update(size_t dt) override;
+	void Update(float dt) override;
 
 	sf::FloatRect GetBBox() const override;
